init new court and extract simple insurance
