+++ utils/src/court.rs
@@ -1,70 +1,64 @@
+use orml_utilities::OrderedSet;
 use parity_scale_codec::{
     Decode,
     Encode,
 };
 use sp_runtime::RuntimeDebug;
-use sp_std::prelude::*;
+use sp_std::{
+    cmp::Ordering,
+    prelude::*,
+};
 
-#[derive(new, PartialEq, Eq, Default, Clone, Encode, Decode, RuntimeDebug)]
-pub struct Dispute<AccountId, Currency, BlockNumber, VoteMetadata, State> {
-    locker: AccountId,
-    locked_funds: Currency,
-    dispute_raiser: AccountId,
-    resolution_metadata: VoteMetadata,
-    state: State,
-    expiry: Option<BlockNumber>,
+#[derive(new, Clone, Copy, Encode, Decode, RuntimeDebug)]
+pub struct Threshold<Rank, Id> {
+    pub rank: Rank,
+    pub id: Id,
 }
 
-impl<
-        AccountId: Clone + PartialEq,
-        Currency: Clone,
-        BlockNumber: Copy,
-        VoteMetadata: Clone,
-        State: Copy,
-    > Dispute<AccountId, Currency, BlockNumber, VoteMetadata, State>
+impl<Rank: Copy + Eq + Ord, Id: Copy + Eq> Ord for Threshold<Rank, Id> {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.rank.cmp(&other.rank)
+    }
+}
+
+impl<Rank: Copy + Eq + Ord, Id: Copy + Eq> Eq for Threshold<Rank, Id> {}
+
+impl<Rank: Copy + Eq + Ord, Id: Copy + Eq> PartialOrd for Threshold<Rank, Id> {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl<Rank: Copy + Eq + Ord, Id: Copy + Eq> PartialEq for Threshold<Rank, Id> {
+    fn eq(&self, other: &Self) -> bool {
+        self.rank == other.rank
+    }
+}
+
+#[derive(PartialEq, Eq, Encode, Decode, RuntimeDebug)]
+/// Sequence of vote metadata for resolving disputes
+pub struct Court<Id, AccountId, Balance, Threshold> {
+    id: Id,
+    controller: Option<AccountId>,
+    // reservation required to trigger vote or appeal to next court
+    bond: Balance,
+    pub vote_seq: OrderedSet<Threshold>,
+}
+
+impl<Id: Copy, AccountId: Clone, Balance: Copy, Threshold: Copy + Ord>
+    Court<Id, AccountId, Balance, Threshold>
 {
-    pub fn locker(&self) -> AccountId {
-        self.locker.clone()
-    }
-    pub fn locked_funds(&self) -> Currency {
-        self.locked_funds.clone()
-    }
-    pub fn dispute_raiser(&self) -> AccountId {
-        self.dispute_raiser.clone()
-    }
-    pub fn can_raise_dispute(&self, who: &AccountId) -> bool {
-        &self.dispute_raiser() == who
-    }
-    pub fn resolution_metadata(&self) -> VoteMetadata {
-        self.resolution_metadata.clone()
-    }
-    pub fn state(&self) -> State {
-        self.state
-    }
-    pub fn expiry(&self) -> Option<BlockNumber> {
-        self.expiry
-    }
-    pub fn set_state(&self, state: State) -> Self {
+    pub fn new(
+        id: Id,
+        controller: Option<AccountId>,
+        bond: Balance,
+        vote_seq: Vec<Threshold>,
+    ) -> Self {
         Self {
-            state,
-            ..self.clone()
+            id,
+            controller,
+            bond,
+            vote_seq: OrderedSet::from(vote_seq),
         }
-    }
-}
-
-#[derive(PartialEq, Eq, Copy, Clone, Encode, Decode, RuntimeDebug)]
-pub enum DisputeState<VoteId> {
-    DisputeNotRaised,
-    /// Dispute raised and vote dispatched without outcome
-    DisputeRaisedAndVoteDispatched(VoteId),
-    /// Outcome and time last checked and outcome updated
-    DisputeRaisedAndAccepted(VoteId),
-    /// Outcome and time last checked and outcome updated
-    DisputeRaisedAndRejected(VoteId),
-}
-
-impl<VoteId> Default for DisputeState<VoteId> {
-    fn default() -> DisputeState<VoteId> {
-        DisputeState::DisputeNotRaised
     }
 }
