+++ utils/src/lib.rs
@@ -13,6 +13,7 @@ pub mod bounty;
 pub mod court;
 pub mod drip;
 pub mod grant;
+pub mod insurance;
 pub mod kickback;
 pub mod meta;
 pub mod moloch;
